'gridded' option works, new Config() class & reworked axis drawing
The new Config() class will become a base class for using to store and create configuration data, however it could be a good idea to move away from this concept and just store the data as object attributes rather than a separate class.

Implemented an efficient algorithm to draw the digit strokes, or if 'gridded' option is selected, a grid.

removed redundant code such as render()

added findCanvas() and findTittle(), to find the tittle div and the canvas tag within the provided html element. It could be a good idea to combine these two methods, to be more dynamic.