Implemented X-Axis & Y-Axis Labels feature, fixed issue with having un-even canvas dimensions not correctly drawing the grid.
The X-Axis and Y-Axis labels are drawn if they are provided as the options, by default they will not be added to the graph object and the padding space for the labels will be re-purposed for the actul graph.

added new function, drawLabels(), before the grid and axis' are drawn, the labels are drawn, since adding the y-axis label requires for the canvas to be rotated and translated.
added new function for retrieving the label font size, getFontSize()
removed 'err' variable, deemed to be redundant
collapsed the functions findCanvas() and findTittle() and setTittlePosition() into one function, findElements(). The function runs over provided node and finds the appropriate html nodes. After constructing a map of the elements, DOM modifications which were where present in the old functions are applied (for simplification).

renamed several variable names which could be deemed to be confusing / miss-representative
created separate variables / properties for the canvas object width and height.

moved the graph calculations into the class constructor for simplification and sharing.